pyproject.toml
[project]
name = "network-ids"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pandas",
    "numpy",
    "scikit-learn",
    "xgboost",
    "torch",
    "kafka-python",
    "mlflow",
    "celery>=5.6.3",
    "redis>=7.4.0",
]
